First use of nqp
Finally gave in and decided to use nqp for one particular thing in the
encoder -- it's a lot faster to check types with:

    if nqp::istype($_, SomeType) {

than with:

    when SomeType {

Just this change at the outermost type checking level made a noticeable
improvement in encoding performance.  Will check if this continues to be
true at deeper levels.